% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjwald_ci-function.R
\name{adjwald_ci}
\alias{adjwald_ci}
\title{Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)}
\usage{
adjwald_ci(.success, .trials, .Z = 1.959964, .return = c("both", "low", "hi"))
}
\arguments{
\item{.success}{the proportion of successes}

\item{.trials}{the total number of trials}

\item{.Z}{The Z score corresponding to the desired level of confidence}

\item{.return}{One of \code{c("both","low","hi")}. Defaults to \code{both}, returning a list with the lower and upper limits.}
}
\value{
a list with the lower and upper confidence limits (%) as numerics
}
\description{
Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)
}
\seealso{
Other point estimate helpers: 
\code{\link{between_diff_ci}()},
\code{\link{tdist_ci}()},
\code{\link{within_diff_ci}()}
}
\concept{point estimate helpers}
