% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_df-function.R
\name{between_df}
\alias{between_df}
\title{Calculate the degrees of freedom for a between-subjects design.}
\usage{
between_df(.sd, .n, .sd2, .n2)
}
\arguments{
\item{.sd}{the standard deviation from sample 1.}

\item{.n}{the size of sample 1.}

\item{.sd2}{the standard deviation from sample 2.}

\item{.n2}{the size of sample 2.}
}
\value{
Returns the between-subjects degrees of freedom
}
\description{
\href{https://g.co/kgs/a7Zyyn}{Sauro and Lewis (2012)} recommend using the Welch-Satterthwaite procedure for computing the degrees of freedom in a between-subjects design.
}
\seealso{
Other t-distribution functions (~ Excel): 
\code{\link{tdist}()},
\code{\link{tinv}()}
}
\concept{t-distribution functions (~ Excel)}
