% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomdist-function.R
\name{binomdist}
\alias{binomdist}
\title{Find a binomial-distribution probability}
\usage{
binomdist(.s, .t, .p, .cumulative = FALSE, ...)
}
\arguments{
\item{.s}{The number of successes. Cannot be less than 0.}

\item{.t}{The number of trials. Must be greater than 0 and greater than or equal to \code{.s}}

\item{.p}{The probability of success on each trial. Must be greater than 0 and less than 1.}

\item{.cumulative}{Defaults to FALSE, returning the probability mass function, which is the probability that there are \code{.s} successes. If TRUE, \code{binomdist} returns the cumulative distribution function, which is the probability that there are at most \code{.s} successes.}

\item{...}{(Optional) Extra arguments passed to \code{stats::pbinom()} or \code{stats::dbinom()}}
}
\value{
Returns the individual term binomial distribution probability.
}
\description{
Similar to Excel's \code{BINOMDIST(number_s,trials,probability_s,cumulative)} function.
\code{binomdist()} uses \code{stats::pbinom()} when \code{.cumulative = TRUE}, and \code{stats::dbinom()} when not (i.e., the default).
}
\concept{binomial-distribution functions (~ Excel)}
