% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace-function.R
\name{laplace}
\alias{laplace}
\title{Computes success rate using Laplace estimate}
\usage{
laplace(.success, .trials)
}
\arguments{
\item{.success}{the total number of successes}

\item{.trials}{the total number of trials}
}
\value{
an estimated success rate (%) as a numeric
}
\description{
Computes success rate using Laplace estimate
}
\seealso{
Other point estimate helpers: 
\code{\link{adjwald_ci}()},
\code{\link{mle}()},
\code{\link{wilson}()}
}
\concept{point estimate helpers}
