% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-function.R
\name{mle}
\alias{mle}
\title{Computes success rate using Maximum Liklihood estimate}
\usage{
mle(.success, .trials)
}
\arguments{
\item{.success}{the total number of successes}

\item{.trials}{the total number of trials}
}
\value{
an estimated success rate (%) as a numeric
}
\description{
Computes success rate using Maximum Liklihood estimate
}
\seealso{
Other point estimate adjusters: 
\code{\link{laplace}()},
\code{\link{wilson}()}
}
\concept{point estimate adjusters}
