% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_mean-function.R
\name{ratings_mean}
\alias{ratings_mean}
\alias{ratings_mean.numeric}
\alias{ratings_mean.data.frame}
\title{Compute a mean and confidence interval for continuously distributed data}
\usage{
ratings_mean(.data, ...)

\method{ratings_mean}{numeric}(.data, ..., .alpha = NULL, .limits = NULL)

\method{ratings_mean}{data.frame}(.data, .var, ..., .alpha = NULL, .limits = NULL)
}
\arguments{
\item{.data}{A vector or long-format data frame of ratings or other continuous data.}

\item{...}{(Optional) The unquoted, comma-separated names of columns containing grouping variables.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, the default value is 0.05, or a 95\% confidence level.}

\item{.limits}{(Optional) If you want to specify the end-points (limits) for the ratings scale, which will ensure that confidence interval values don't exceed the upper and lwoer bounds, you can supply a numeric vector of length two,indicating the limits (e.g., \code{.limits = c(1,7)}).}

\item{.var}{(Optional) The unquoted name of the data frame column containing the values to use in the computation.}
}
\value{
a tibble with one or more means
}
\description{
For ratings data and other continuously distributed variables, \code{ratings_mean()} returns a mean, along with information about the confidence interval (based on the T distribution). You can optionally include one or more grouping variables to compute means by groups; modify the alpha level to adjust the confidence intervals; and specific scale limits so that the output values have upper- and lower-bounds.
}
\examples{
# Compare the difference between:
 ratings_mean(c(1,8,8))

 #and:

  ratings_mean(c(1,8,8), .limits = c(1,8))
 @export

}
\concept{means with confidence intervals}
