% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_vs_bench-function.R
\name{ratings_vs_bench}
\alias{ratings_vs_bench}
\alias{ratings_vs_bench.numeric}
\alias{ratings_vs_bench.data.frame}
\title{Compare ratings to a benchmark}
\usage{
ratings_vs_bench(.x, ...)

\method{ratings_vs_bench}{numeric}(
  .x,
  .sd,
  .n,
  .m,
  ...,
  .alt = c("greater", "less", "twotailed"),
  .alpha = 0.05
)

\method{ratings_vs_bench}{data.frame}(.x, .var, .m, ..., .alt = "greater", .alpha = 0.05)
}
\arguments{
\item{.x}{A single numeric value, a vector of values, or a long-format data frame with a named column of numeric data corresponding to ratings or another continuous metric. See Details.}

\item{...}{(Optional) If \code{.x} is a long-format data frame, you can pass the name of one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.}

\item{.sd}{If \code{.x} is a single numeric value, \code{.sd} should be a single numeric value indicating the sample standard deviation. See Details}

\item{.n}{If \code{.x} is a single numeric value, \code{.n} should be a single numeric value indicating the sample size. See Details.}

\item{.m}{The test (benchmark) mean.}

\item{.alt}{(Optional) For test alternatives, one of \code{c("greater","less","twotailed")}. Defaults to "greater" for a one-sided test.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the significance level to be used. Defaults to \code{.alpha = 0.05}. To set a different significance level, the argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results.}

\item{.var}{If \code{.x} is a long-format data frame, the (unquoted) name of a data frame column containing task ratings (or values from another continuous metric).}
}
\value{
A tibble with data summaries and test results
}
\description{
\code{ratings_vs_bench()} tests a sample mean of continuous data (e.g., ratings)  against a given benchmark using a one-sample t-test.
}
\details{
\code{ratings_vs_bench} assumes that you want to test the hypothesis that the observed outcome \emph{exceeds} the benchmark, and therefore, defaults to a one-tailed test. This means that setting \code{.alpha = 0.05} (the default) produces a 90\% confidence interval.

\itemize{
 \item If \code{.x} is a single numeric value representing the sample mean, provide the sample standard deviation and sample size to \code{.sd} and \code{.n}, respectively.
 \item If \code{.x} is a numeric vector of ratings values, you should only specify the test (benchmark) mean.
 \item If \code{.x} is a data frame, \code{.var} should be the unquoted name of the column containing the ratings values.
 \item If you're passing a data frame to \code{.x}, you can optionally pass one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.
  \item You can choose from among the test alternatives \code{c("greater","less","twotailed")} by providing one of the options to the \code{.alt} argument: e.g., \code{.alt = "twotailed"}. Defaults to "greater" for a one-sided test.
 \item You can modify the alpha level to adjust confidence intervals by including \code{.alpha} as a named argument and providing a numeric value: e.g., \code{.aplha = 0.001}.
}

Note that \code{NAs} are automatically dropped in all calculations.
}
\examples{
# Comparing a sample ratings mean of 79 (sd = 20) from 160 users
# against a benchmark mean of 75
ratings_vs_bench(.x=79,.sd=20,.n=160,.m=75)

# Comparing values from a data frame against a benchmark of 70.
.ux_data <-
 data.frame(
  "id" = rep(seq(1,10,1),2),
  "task" = c(rep(1,10),rep(2,10)),
  "ratings"  = runif(20,0,100))

ratings_vs_bench(.ux_data, ratings, .m=70, task)

}
\seealso{
Other benchmark comparison stats: 
\code{\link{success_vs_bench}()},
\code{\link{time_vs_bench}()}
}
\concept{benchmark comparison stats}
