% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdist-function.R
\name{tdist}
\alias{tdist}
\title{Find a T-distribution probability}
\usage{
tdist(.p, .df, .tail = 2)
}
\arguments{
\item{.p}{The numeric probability associated with the two-tailed T-distribution. Must be greater than 0 and less than 1.}

\item{.df}{The number of degrees of freedom.Must be greater than 0 and less than 1.}

\item{.tail}{One of \code{1} or \code{2} (Default) indicating whether to return the one-tailed or two-tailed value.}
}
\value{
Returns the probability for the Student T-distribution when .x is given as the value of T.
}
\description{
Calculates the one- or two-tailed probability density for Student's T Distribution

Similar to Excel's \code{TDIST(abs(x),df,tails)} function.
}
\seealso{
Other t-distribution functions (~ Excel): 
\code{\link{tinv}()}
}
\concept{t-distribution functions (~ Excel)}
