% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdist_ci-function.R
\name{tdist_ci}
\alias{tdist_ci}
\title{Computes confidence intervals using the T distribution}
\usage{
tdist_ci(.mean, .sd, .n, .t_crit, .return = c("both", "low", "hi"))
}
\arguments{
\item{.mean}{the sample mean for which the CI is being estimated}

\item{.sd}{The sample standard deviation}

\item{.n}{The sample size}

\item{.t_crit}{the critical T value}

\item{.return}{One of \code{c("both","low","hi")}. Defaults to \code{both}, returning a list with the lower and upper limits}
}
\value{
a list with the lower and upper confidence limits, or
}
\description{
Computes confidence intervals using the T distribution
}
\seealso{
Other point estimate helpers: 
\code{\link{adjwald_ci}()}
}
\concept{point estimate helpers}
