% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means_between-function.R
\name{test_means_between}
\alias{test_means_between}
\alias{ab_test_between}
\alias{test_means_between.numeric}
\alias{test_means_between.data.frame}
\title{Test two means against each other for a "between-subjects" design}
\usage{
test_means_between(.x, ...)

ab_test_between(.x, ...)

\method{test_means_between}{numeric}(.x, .sd, .n, .x2, .sd2, .n2, ..., .alpha = 0.05)

\method{test_means_between}{data.frame}(.x, .var1, .var2, ..., .alpha = 0.05)
}
\arguments{
\item{.x}{A single numeric value (i.e., the mean from sample 1) or a data frame with named columns of numeric data corresponding to the variables being tested. See Details.}

\item{...}{(Optional) If \code{.x} is a long-format data frame with a separate unique column of data for each sample, you can pass the name of one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.}

\item{.sd}{If \code{.x} is a single numeric value representing the mean of the variable for sample 1, \code{.sd} should be a single numeric value representing the standard deviation of the variable for sample 1. See Details.}

\item{.n}{If \code{.x} is a single numeric value representing the mean of the variable for sample 1, \code{.sd} should be a single numeric value representing the total number of users in  sample 1. See Details.}

\item{.x2}{If \code{.x} is a single numeric value representing the mean of the variable for sample 1,\code{.x2} should be a single numeric value representing the mean from sample 2.}

\item{.sd2}{If \code{.x} is a single numeric value representing the mean of the variable for sample 1, \code{.sd2} should be a single numeric value representing the standard deviation of the variable for sample 2. See Details.}

\item{.n2}{If \code{.x} is a single numeric value representing the mean of the variable for sample 1, \code{.sd} should be a single numeric value representing the total number of users in  sample 2. See Details.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the significance level to be used. Defaults to \code{.alpha = 0.05}. To set a different significance level, the argument must be named (i.e., \code{.alpha = 0.001}) or else the function may yield unexpected results.}

\item{.var1}{If \code{.x} is a long-format data frame, the (unquoted) name of a data frame column containing the values representing the first sample.}

\item{.var2}{If \code{.x} is a long-format data frame, the (unquoted) name of a data frame column containing the values representing the second sample.}
}
\value{
A tibble with the results of a test of two means, with confidence interval information, and other information.
}
\description{
\code{test_means_between()} compares two sample means obtained from two different groups of users  and performs a two-sided statistical test to determine if the means differ from each other.

\code{test_means_between}, \code{ab_test_between}, and \code{test_times_between} are synonyms.
}
\details{
\itemize{
 \item If \code{.x} is a single numeric value representing the mean of the the variable for sample 1, use \code{.sd} and \code{.n} to provide the standard deviation and sample size. Use \code{.x2,.sd2,.n2} to provide the equivalent information for sample 2.
 \item If \code{.x} is a data frame with unique variable columns for each sample,\code{.var1} and \code{.var2} should be the unquoted names of the columns containing the values for sample 1 and sample 2
 \item If \code{.x} is a data frame with only one variable column for both samples, pass the unquoted name of the column containing the values to compute means over to \code{.var1}, set \code{.var2 = NULL}, and use the grouping argument (\code{...}) to provide the name of the column that contains the sample id, which should only contain two unique values. Note that if you use this option, you cannot pass other grouping variables.
 \item You can modify the alpha level to adjust confidence intervals by including \code{.alpha} as a named argument and providing a numeric value: e.g., \code{.aplha = 0.001}.
 \item If you're passing a data frame to \code{.x}, you can optionally pass one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.
}

Note that \code{NAs} are automatically dropped in all calculations.
}
\examples{
# When you already have summary stats:

test_means_between(.x = 45,.sd = 9.2,.n = 18,.x2 = 48, .sd2 = 8.5, .n2 = 17, .alpha = 0.01)


# If you have a data frame with a separate column of observations
# for each sample:

.ux_data<-
  data.frame("sample1_ratings" = runif(20,20,50),
             "sample2_ratings" = c(runif(15,20,50),rep(NA,times=5)))

test_means_between(.x = .ux_data, .var1 = sample1_ratings, .var2 = sample2_ratings, .alpha = 0.10)


}
\seealso{
Other tests comparing means: 
\code{\link{test_means_within}()}
}
\concept{tests comparing means}
