% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_vs_benchmark-function.R
\name{time_vs_bench}
\alias{time_vs_bench}
\alias{time_vs_bench.numeric}
\alias{time_vs_bench.data.frame}
\title{Compare sample time to a benchmark}
\usage{
time_vs_bench(.x, ...)

\method{time_vs_bench}{numeric}(.x, .m, .sd, .n, ..., .alpha = 0.1)

\method{time_vs_bench}{data.frame}(.x, .m, .var, ..., .alpha = 0.1)
}
\arguments{
\item{.x}{The number of observed successes.}

\item{...}{(Optional) The unquoted, comma-separated names of columns containing grouping variables.}

\item{.m}{The benchmark mean.}

\item{.sd}{The sample standard deviation.}

\item{.n}{The total number of users.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, the default value is 0.05, corresponding to a 90\% confidence level for a one-sided test.}

\item{.var}{The (unquoted) name of the column containing the time data.
.uxdata <-
data.frame(
 "id" = rep(seq(1,10,1),2),
 "group" = rep(c("A","B"),10),
 "task" = c(rep(1,10),rep(2,10)),
 "time"  = runif(20,0,1000)
)

time_stats(.uxdata, .var=time, group, task, .alpha = 0.001)}
}
\value{
A tibble with test results
}
\description{
Compare sample time to a benchmark
}
\seealso{
Other benchmark comparison stats: 
\code{\link{ratings_vs_bench}()},
\code{\link{success_vs_bench}()}
}
\concept{benchmark comparison stats}
