% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinv-function.R
\name{tinv}
\alias{tinv}
\title{Calculate the inverse of the two-tailed Student's T Distribution}
\usage{
tinv(.a, .df, .tail = 2, ...)
}
\arguments{
\item{.a}{The alpha level, or the numeric probability associated with the T-distribution. Must be greater than 0 and less than 1.}

\item{.df}{The number of degrees of freedom. Must be greater than 0 and less than 1.}

\item{.tail}{One of \code{1} or \code{2} (Default) indicating whether to return the one-tailed or two-tailed value.}

\item{...}{(Optional) Named arguments to be passed to \code{stats::qt()}.}
}
\value{
Returns the (absolute) T-critical value.
}
\description{
Equivalent to Excel's \code{TINV(x,df)} function.
}
\seealso{
Other t-distribution functions (~ Excel): 
\code{\link{between_df}()},
\code{\link{tdist}()}
}
\concept{t-distribution functions (~ Excel)}
