% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_mean_test-function.R
\name{two_mean_test}
\alias{two_mean_test}
\alias{ab_test}
\alias{two_mean_test.numeric}
\alias{two_mean_test.data.frame}
\title{Test two means against each other}
\usage{
two_mean_test(.x, ...)

ab_test(.x, ...)

\method{two_mean_test}{numeric}(.x, .sd, .n, ..., .alpha = 0.05)

\method{two_mean_test}{data.frame}(.x, .var1, .var2, ..., .alpha = 0.05)
}
\arguments{
\item{.x}{A single numeric value or a long-format data frame with named columns of numeric data corresponding to the variables being tested. See Details.}

\item{...}{(Optional) If \code{.x} is a long-format data frame, you can pass the name of one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.}

\item{.sd}{If \code{.x} is a single numeric value representing the mean of the variable differences, \code{.sd} should be a single numeric value representing the standard deviation of the diferences. See Details.}

\item{.n}{If \code{.x} is a single numeric value representing the mean of the variable differences, \code{.sd} should be a single numeric value representing the total number of trials. See Details.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the significance level to be used. Defaults to \code{.alpha = 0.05}. To set a different significance level, the argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results.}

\item{.var1}{If \code{.x} is a long-format data frame, the (unquoted) name of a data frame column containing the values for computing the first mean.}

\item{.var2}{If \code{.x} is a long-format data frame, the (unquoted) name of a data frame column containing the values for computing the first mean.}
}
\value{
A tibble with the results of a test of two means, with confidence interval information, and other information.
}
\description{
\code{two_mean_test()} compares to sample means and performs a two-sided statistical test to determine if the means differ from each other.

\code{two_mean_test}, \code{ab_test}, and \code{two_times_test} are synonyms.
}
\details{
\itemize{
 \item If \code{.x} is a single numeric value representing the mean of differences from the sample scores, \code{.sd} should be a single numeric value representing the standard deviation of the difference scores.
 \item If \code{.x} is a single numeric value representing the mean of differences from the sample scores, \code{.n} should be a single numeric value representing the number of pairs that derived the difference scores.
 \item If \code{.x} is a data frame, \code{.var1} and \code{.var2} should be the unquoted names of the columns containing the raw scores obtained.
 \item You can modify the alpha level to adjust confidence intervals by including \code{.alpha} as a named argument and providing a numeric value: e.g., \code{.aplha = 0.001}.
 \item If you're passing a data frame to \code{.x}, you can optionally pass one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.
}

Note that \code{NAs} are automatically dropped in all calculations.
}
\examples{
# When you have computed difference metrics in advance:
two_mean_test(45,9.2,18)

.ux_data <-
 data.frame(
  "id" = rep(seq(1,20,1),2),
  "task" = c(rep(1,20),rep(2,20)),
  "score1"  = sample(1:7,40,TRUE),
  "score2"  = sample(1.5:7.5,40,TRUE)
  )

two_mean_test(.ux_data, score1, score2, task, .alpha=0.1)
}
\concept{tests comparing means}
