% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_times_test-function.R
\name{two_times_test}
\alias{two_times_test}
\title{Test two means against each other}
\usage{
two_times_test(.x, ...)
}
\arguments{
\item{.x}{A single numeric value or a long-format data frame with named columns of numeric data corresponding to the variables being tested. See Details.}

\item{...}{(Optional) If \code{.x} is a long-format data frame, you can pass the name of one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.}
}
\value{
A tibble with the results of a test of two means, with confidence interval information, and other information.
}
\description{
\code{two_mean_test()} compares to sample means and performs a two-sided statistical test to determine if the means differ from each other.

\code{two_mean_test}, \code{ab_test}, and \code{two_times_test} are synonyms.
}
\details{
\itemize{
 \item If \code{.x} is a single numeric value representing the mean of differences from the sample scores, \code{.sd} should be a single numeric value representing the standard deviation of the difference scores.
 \item If \code{.x} is a single numeric value representing the mean of differences from the sample scores, \code{.n} should be a single numeric value representing the number of pairs that derived the difference scores.
 \item If \code{.x} is a data frame, \code{.var1} and \code{.var2} should be the unquoted names of the columns containing the raw scores obtained.
 \item You can modify the alpha level to adjust confidence intervals by including \code{.alpha} as a named argument and providing a numeric value: e.g., \code{.aplha = 0.001}.
 \item If you're passing a data frame to \code{.x}, you can optionally pass one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.
}

Note that \code{NAs} are automatically dropped in all calculations.
}
\examples{
# When you have computed difference metrics in advance:
two_mean_test(45,9.2,18)

.ux_data <-
 data.frame(
  "id" = rep(seq(1,20,1),2),
  "task" = c(rep(1,20),rep(2,20)),
  "score1"  = sample(1:7,40,TRUE),
  "score2"  = sample(1.5:7.5,40,TRUE)
  )

two_mean_test(.ux_data, score1, score2, task, .alpha=0.1)
}
