% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_mean-function.R
\name{ux_mean}
\alias{ux_mean}
\alias{ux_mean.default}
\alias{ux_mean.numeric}
\alias{ux_mean.data.frame}
\title{Returns a mean with confidence intervals based on the t-distribution.}
\usage{
ux_mean(.data, ...)

\method{ux_mean}{default}(.data, ...)

\method{ux_mean}{numeric}(.data, ..., .alpha = NULL)

\method{ux_mean}{data.frame}(.data, .var, ..., .alpha = NULL)
}
\arguments{
\item{.data}{A vector or data frame of values (ratings)}

\item{...}{(Optional) The unquoted, comma-separated names of columns containing grouping variables.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} <1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, it defaults to 0.05.}

\item{.var}{(Optional) The unquoted name of the data frame column containing the values to use in the computation.}
}
\value{
a tibble with one or more means
}
\description{
Returns a mean with confidence intervals based on the t-distribution.
}
\concept{means with confidence intervals}
