% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_success_rate-function.R
\name{ux_success_rate}
\alias{ux_success_rate}
\alias{ux_success_rate.default}
\alias{ux_success_rate.numeric}
\alias{ux_success_rate.data.frame}
\title{Computes estimated success rate with confidence intervals for small samples, following Sauro and Lewis (2012)}
\usage{
ux_success_rate(.x, .y, ...)

\method{ux_success_rate}{default}(.x, .y, ...)

\method{ux_success_rate}{numeric}(.x, .y, ..., .alpha = NULL)

\method{ux_success_rate}{data.frame}(.x, .y, ..., .alpha = NULL)
}
\arguments{
\item{.x}{You can pass an integer (>0) indicating the total number of successes, and provide the total number of trials  to \code{.y} (where the value provided to \code{.y} > \code{.y}). Or you can pass a data frame containing trial data.}

\item{.y}{If \code{.x} is an integer representing the total number of successes, \code{.y} should be an integer indicating the total number of trials. Or, if \code{.x} is a data frame, provide the unquoted name of the column containing the success data (as 1s and 0s) to \code{.y}.}

\item{...}{(Optional) If \code{.x} is a data frame, you can pass the name of grouping variables here as unquote, comma-separated column names.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} <1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, it defaults to 0.05.}
}
\value{
a tibble with
}
\description{
Computes estimated success rate with confidence intervals for small samples, following Sauro and Lewis (2012)
}
\concept{success rate estimators}
