% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson-function.R
\name{wilson}
\alias{wilson}
\title{Computes success rate using Wilson estimate}
\usage{
wilson(.success, .trials, .Z = 1.959964)
}
\arguments{
\item{.success}{the total number of successes}

\item{.trials}{the total number of trials}

\item{.Z}{The Z score corresponding to the desired level of confidence}
}
\value{
an estimated success rate (%) as a numeric
}
\description{
Computes success rate using Wilson estimate
}
\seealso{
Other point estimate helpers: 
\code{\link{adjwald_ci}()},
\code{\link{laplace}()},
\code{\link{mle}()}
}
\concept{point estimate helpers}
