% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_diff_ci-function.R
\name{within_diff_ci}
\alias{within_diff_ci}
\title{Computes confidence intervals on the difference of means for within-subjects tests}
\usage{
within_diff_ci(.mean, .sd, .n, .alpha, .return = c("both", "low", "hi"))
}
\arguments{
\item{.mean}{the sample mean for which the CI is being estimated}

\item{.sd}{The sample standard deviation}

\item{.n}{The sample size}

\item{.alpha}{the critical T value}

\item{.return}{One of \code{c("both","low","hi")}. Defaults to \code{both}, returning a list with the lower and upper limits}
}
\value{
a list with the lower and upper confidence limits, or
}
\description{
Computes confidence intervals on the difference of means for within-subjects tests
}
\seealso{
Other point estimate helpers: 
\code{\link{adjwald_ci}()},
\code{\link{tdist_ci}()}
}
\concept{point estimate helpers}
