% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjwald_ci-function.R
\name{adjwald_ci}
\alias{adjwald_ci}
\title{Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)}
\usage{
adjwald_ci(.success, .trials, .Z = 1.959964)
}
\arguments{
\item{.success}{the proportion of successes}

\item{.trials}{the total number of trials}

\item{.Z}{The Z score corresponding to the desired level of confidence}
}
\value{
a list with the lower and upper confidence limits (%) as numerics
}
\description{
Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)
}
\seealso{
Other point estimate helpers: 
\code{\link{tdist_ci}()}
}
\concept{point estimate helpers}
