% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_stats-function.R
\name{ratings_stats}
\alias{ratings_stats}
\alias{ratings_stats.numeric}
\alias{ratings_stats.data.frame}
\title{Compute stats for ratings or continuously distributed data}
\usage{
ratings_stats(.x, ...)

\method{ratings_stats}{numeric}(.x, ..., .alpha = 0.05, .limits = NULL)

\method{ratings_stats}{data.frame}(.x, .var, ..., .alpha = 0.05, .limits = NULL)
}
\arguments{
\item{.x}{A vector of values, or a long-format data frame with a named column containing numeric ratings data.}

\item{...}{(Optional) If \code{.x} is a long-format data frame, you can pass the name of one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the significance level to be used. Defaults to \code{.alpha = 0.05}. To set a different significance level, the argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results.}

\item{.limits}{(Optional) If you want to specify the end-points (limits) for the ratings scale, which will ensure that confidence interval values don't exceed the upper and lower bounds, you can supply a numeric vector of length two,indicating the limits (e.g., \code{.limits = c(1,7)}).}

\item{.var}{If \code{.x} is a data frame, the unquoted name of the data frame column containing the values to use in the computations.}
}
\value{
A tibble with one or more means, confidence interval information, and other information.
}
\description{
For ratings data and other continuously distributed variables, \code{ratings_stats()} returns means; information about confidence intervals (based on the T distribution); standard deviations; medians; and other details.
}
\details{
\itemize{
 \item You can modify the alpha level to adjust confidence intervals by including \code{.alpha} as a named argument and providing a numeric value: e.g., \code{.aplha = 0.001}.
 \item You can specific scale limits so that the output values have upper- and lower-bounds by including \code{.limits} and providing a numeric vector of length 2: e.g., \code{.limits = c(1.5,6.5)}.
 \item If you're passing a data frame to \code{.x}, you can optionally pass one or more grouping variables as unquoted, comma-separated column names (without naming the \code{...} argument) to compute stats by groups.
}

Note that \code{NAs} are automatically dropped in all calculations.
}
\examples{

# Compare the difference between the output of:
ratings_stats(c(1,8,8)) # and:

ratings_stats(c(1,8,8), .limits = c(1,8))

.ux_data <-
 data.frame(
  "id" = rep(seq(1,10,1),2),
  "task" = c(rep(1,10),rep(2,10)),
  "easiness"  = sample(1:7,20,replace=TRUE))

ratings_stats(.ux_data, easiness,task,.alpha=0.01,.limits=c(1,7))

}
\seealso{
Other descriptive stats for UX measures: 
\code{\link{problem_stats}()},
\code{\link{success_stats}()},
\code{\link{time_stats}()}
}
\concept{descriptive stats for UX measures}
