% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_vs_bench.R
\name{ratings_vs_bench}
\alias{ratings_vs_bench}
\alias{ratings_vs_bench.numeric}
\alias{ratings_vs_bench.data.frame}
\title{Compare ratings to a benchmark}
\usage{
ratings_vs_bench(.x, ...)

\method{ratings_vs_bench}{numeric}(.x, .m, .sd, .n, ..., .alpha = 0.1)

\method{ratings_vs_bench}{data.frame}(.x, .m, .var, ..., .alpha = 0.1)
}
\arguments{
\item{.x}{The sample mean.}

\item{...}{(Optional) The unquoted, comma-separated names of columns containing grouping variables.}

\item{.m}{The benchmark mean.}

\item{.sd}{The sample standard deviation.}

\item{.n}{The sample size}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, the default value is 0.05, corresponding to a 90\% confidence level for a one-sided test.}

\item{.var}{The benchmark mean.}
}
\value{
A tibble with test results
}
\description{
Compare ratings to a benchmark
}
\examples{

.uxdata <-
data.frame(
 "id" = rep(seq(1,10,1),2),
 "group" = rep(c("A","B"),10),
 "task" = c(rep(1,10),rep(2,10)),
 "ratings"  = runif(20,0,100)
)

ratings_vs_bench(.uxdata, .m=75, .var=ratings,group)
}
\seealso{
Other means with confidence intervals: 
\code{\link{ratings_stats}()},
\code{\link{success_vs_bench}()},
\code{\link{time_stats}()}
}
\concept{means with confidence intervals}
