% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_vs_bench-function.R
\name{success_vs_bench}
\alias{success_vs_bench}
\alias{success_vs_bench.numeric}
\alias{success_vs_bench.data.frame}
\title{Compare a success rate to a benchmark}
\usage{
success_vs_bench(.x, ...)

\method{success_vs_bench}{numeric}(
  .x,
  .n,
  .p,
  ...,
  .alt = c("greater", "less", "twotailed"),
  .alpha = 0.05
)

\method{success_vs_bench}{data.frame}(
  .x,
  .n,
  .p,
  ...,
  .alt = c("greater", "less", "twotailed"),
  .alpha = 0.05
)
}
\arguments{
\item{.x}{The number of observed successes.}

\item{...}{(Optional) The unquoted, comma-separated names of columns containing grouping variables.}

\item{.n}{The total number of users.}

\item{.p}{The test proportion (must be a numeric between 0-1).}

\item{.alt}{For alternatives, one of \code{c("greater","less","twotailed")}. Defaults to "greater" for a one-sided test.}

\item{.alpha}{(Optional) A positive number (where 0 < \code{.alpha} < 1) specifying the desired confidence level to be used. The argument must be named (i.e., \code{.alpha=0.001}) or else the function may yield unexpected results. If the argument is omitted, the default value is 0.05, corresponding to a 90\% confidence level for a one-sided test.}
}
\value{
A tibble with test results
}
\description{
Compare a success rate to a benchmark
}
\examples{

.uxdata <-
data.frame(
 "id" = rep(seq(1,10,1),2),
 "group" = rep(c("A","B"),10),
 "task" = c(rep(1,10),rep(2,10)),
 "complete"  = sample(0:1,20,replace=TRUE,prob = c(.3,.65))
)

success_vs_bench(.uxdata, complete, .p=0.7,group)
}
\seealso{
Other benchmark comparison stats: 
\code{\link{ratings_vs_bench}()},
\code{\link{time_vs_bench}()}
}
\concept{benchmark comparison stats}
